all shells working
