+++ packages/base-shell/vite.config.js
@@ -11,6 +11,7 @@ export default defineConfig({
     },
   },
   build: {
+    minify: false,
     lib: {
       entry: [
         resolve(__dirname, "src", "index.js"),
