+++ packages/rmw-shell/src/providers/Firebase/Paths/index.jsx
@@ -1,8 +1,8 @@
-import { useContext } from 'react'
-import Context from './Context'
-export { default as withPaths } from './with.js'
-export { default } from './Provider.js'
+import { useContext } from "react";
+import Context from "./Context";
+export { default as withPaths } from "./with";
+export { default } from "./Provider";
 
 export function usePaths() {
-  return useContext(Context)
+  return useContext(Context);
 }
